Started support for several boxes, to make it possible to implement GEMC if needed later. By default, only one box will be used in the first place
